refactor(rest): extract admin audit-run routes into routes/admin_audit_runs.py
Moves the 11 admin audit-run routes (POST/GET/DELETE audit/runs and its
per-run cancel/re-dispatch/rerun-cell/validate/cells/parity actions,
GET audit/active, GET audit/cell-history) out of create_app's closure
into routes/admin_audit_runs.py.

Lifts WASM_POOL, audit_cells_for_files, audit_run_list_cells,
audit_dispatch and audit_dispatch_wasm out of the closure as
RestContext-taking functions — they are also called from the audit
scheduler tick, the auto-validate poller, and the admin audit-schedules
fire-now route, all still inside create_app. create_app keeps the old
underscore-prefixed names bound to them for those callers.
